+++ R/build-reference.R
@@ -78,6 +78,8 @@
 #' * `lacks_concepts(c("concept1", "concept2"))` to select all topics
 #'    without those concepts. This is useful to capture topics not otherwise
 #'    captured by `has_concepts()`.
+#' * Topics from other installed packages, e.g. `rlang::is_installed()` (function name)
+#'  or `sass::font_face` (topic name).
 #'
 #' All functions (except for `has_keywords()`) automatically exclude internal
 #' topics (i.e. those with `\keyword{internal}`). You can choose to include
@@ -218,11 +220,17 @@ build_reference_index <- function(pkg = ".") {
     dir_copy_to(pkg, src_icons, dst_icons)
   }
 
-  invisible(render_page(
+  render_page(
     pkg, "reference-index",
     data = data_reference_index(pkg),
     path = "reference/index.html"
-  ))
+  )
+
+  html <- xml2::read_html(file.path(pkg$dst, "reference/index.html"))
+  tweak_all_links(html, pkg = pkg)
+  xml2::write_html(html, file.path(pkg$dst, "reference/index.html"))
+
+  invisible()
 }